function/requirements.txt
google-cloud-dlp
google-cloud-secret-manager
requests